requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
djangorestframework==3.13.1
flake8==4.0.1
mccabe==0.6.1
Pillow==9.0.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
standalone==1.0.1
